Add helpers for eventData
